fix keyEvent leaks (#560)
